+++ R/learner_lightgbm_classif_lightgbm.R
@@ -504,6 +504,17 @@ LearnerClassifLightGBM = R6Class("LearnerClassifLightGBM",
       map(self$model$record_evals$test, function(metric) {
         metric$eval[[length(metric$eval)]]
       })
+    },
+
+    .extract_best_valid_scores = function() {
+      # the best iteration is only tracked when early stopping is enabled
+      if (is.null(self$model$record_evals$test) || is.null(self$state$param_vals$early_stopping_rounds)) {
+        return(named_list())
+      }
+
+      map(self$model$record_evals$test, function(metric) {
+        metric$eval[[self$model$best_iter]]
+      })
     }
   ),
 
@@ -513,6 +524,14 @@ LearnerClassifLightGBM = R6Class("LearnerClassifLightGBM",
     #' Extracted from `model$evaluation_log`
     internal_valid_scores = function() {
       self$state$internal_valid_scores
+    },
+
+    #' @field best_valid_scores
+    #' The validation scores for all metrics at the best iteration (`model$best_iter`), i.e. the iteration
+    #' that is also reported via `$internal_tuned_values`.
+    #' Only tracked when `early_stopping_rounds` is set, otherwise an empty list.
+    best_valid_scores = function() {
+      self$state$best_valid_scores
     },
 
     #' @field internal_tuned_values